requirements.txt
# local package
-e .

# external requirements

flake8>=3.7.9
geopandas>=0.8.0
kneed>=0.6.0
numpy>=1.18.5
pandas>=1.0.1
plotly>=4.9.0
rpy2>=3.3.5
python-dotenv>=0.14.0
scikit-learn>=0.23.1
scipy>=1.4.1

